# cmd/api/users.go
package main

import (
	"context"
	"errors"
	"net/http"
	"strconv"

	"github.com/go-chi/chi/v5"
	"github.com/sawmeraw/gogo/internal/store"
)

type UserKey string

var userCtx UserKey = "user"

type UserActionPayload struct {
	UserID int64 `json:"user_id"`
}

// GetUser godoc
//
//	@Summary		Fetches a user profile
//	@Description	Fetches a user profile by ID
//	@Tags			users
//	@Accept			json
//	@Product		json
//	@Param			id	path		int	true	"User ID"
//	@Success		200	{object}	store.User
//	@Failure		400	{object}	error
//	@Failure		404	{object}	error
//	@Failure		500	{object}	error
//	@Security		ApiKeyAuth
//	@Router			/users/{id} [get]
func (app *application) getUserHandler(w http.ResponseWriter, r *http.Request) {
	user := getUserFromContext(r)
	if err := app.jsonResponse(w, http.StatusOK, user); err != nil {
		app.statusInternalServerError(w, r, err)
	}
}

// FollowUser godoc
//
//	@Summary		Follow a user
//	@Description	Follows a user by user ID
//	@Tags			users
//	@Accept			json
//	@Product		json
//	@Param			userID	path		int					true	"User ID"
//	@Param			payload	body		UserActionPayload	true	"Follow options"
//	@Success		204		{string}	string				"User followed"
//	@Failure		400		{object}	error				"User payload mising"
//	@Failure		404		{object}	error				"User not found"
//	@Failure		500		{object}	error				"Internal Server Error"
//	@Security		ApiKeyAuth
//	@Router			/users/{userID}/follow [put]
func (app *application) followUserHandler(w http.ResponseWriter, r *http.Request) {
	contextUser := getUserFromContext(r)

	//revert back to auth userID from ctx
	var payload UserActionPayload
	if err := readJSON(w, r, &payload); err != nil {
		app.badRequestResponse(w, r, err)
		return
	}

	//make sure that both ids arent the same
	if contextUser.ID == payload.UserID {
		app.badRequestResponse(w, r, store.ErrNotAllowed)
		return
	}

	ctx := r.Context()
	if err := app.store.Followers.Follow(ctx, contextUser.ID, payload.UserID); err != nil {
		switch err {
		case store.ErrConflict:
			app.conflictResponse(w, r, err)
			return
		default:
			app.statusInternalServerError(w, r, err)
			return
		}
	}

	statusResponse(w, http.StatusNoContent)
}

func (app *application) activateUserHandler(w http.ResponseWriter, r *http.Request) {
	token := chi.URLParam(r, "token")

	err := app.store.Users.Activate(r.Context(), token)
	if err != nil {
		switch err {
		case store.ErrNotFound:
			app.badRequestResponse(w, r, err)
		default:
			app.statusInternalServerError(w, r, err)
		}
		return
	}

	if err := app.jsonResponse(w, http.StatusNoContent, ""); err != nil {
		app.statusInternalServerError(w, r, err)
	}
}

// UnfollowUser godoc
//
//	@Summary		Unfollow a user
//	@Description	Unfollows a user by user ID
//	@Tags			users
//	@Accept			json
//	@Product		json
//	@Param			userID	path		int					true	"User ID"
//	@Param			payload	body		UserActionPayload	true	"Follow options"
//	@Success		204		{string}	string				"User unfollowed"
//	@Failure		400		{object}	error				"User payload mising"
//	@Failure		404		{object}	error				"User not found"
//	@Failure		500		{object}	error				"Internal Server Error"
//	@Security		ApiKeyAuth
//	@Router			/users/{userID}/unfollow [put]
func (app *application) unfollowUserHandler(w http.ResponseWriter, r *http.Request) {
	contextUser := getUserFromContext(r)

	payload := UserActionPayload{}
	if err := readJSON(w, r, &payload); err != nil {
		app.badRequestResponse(w, r, err)
		return
	}

	//make sure that both ids arent the same
	if contextUser.ID == payload.UserID {
		app.badRequestResponse(w, r, store.ErrNotAllowed)
		return
	}

	ctx := r.Context()

	if err := app.store.Followers.UnFollow(ctx, contextUser.ID, payload.UserID); err != nil {
		app.statusInternalServerError(w, r, err)
		return
	}

	statusResponse(w, http.StatusNoContent)
}

func (app *application) usersContextMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		userID, err := strconv.ParseInt(chi.URLParam(r, "userID"), 10, 64)
		if err != nil {
			app.badRequestResponse(w, r, err)
			return
		}

		if userID <= 0 {
			app.badRequestResponse(w, r, store.ErrInvalidArgument)
			return
		}

		ctx := r.Context()
		user, err := app.store.Users.GetByID(ctx, userID)
		if err != nil {
			switch {
			case errors.Is(err, store.ErrNotFound):
				app.notFoundResponse(w, r, err)

			default:
				app.statusInternalServerError(w, r, err)

			}
			return

		}

		ctx = context.WithValue(ctx, userCtx, user)
		next.ServeHTTP(w, r.WithContext(ctx))

	})
}

func getUserFromContext(r *http.Request) *store.User {
	user, _ := r.Context().Value(userCtx).(*store.User)
	return user
}